# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinVault API
  description: Banking simulator API for test automation practice
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
...
  /transfers:
    get:
      tags: [Transfers]
      security: [{ bearerAuth: [] }]
    post:
      tags: [Transfers]
      security: [{ bearerAuth: [] }]
  /transfers/scheduled:
    get:
      tags: [Transfers]
      security: [{ bearerAuth: [] }]
  /transfers/{id}:
    get:
      tags: [Transfers]
      security: [{ bearerAuth: [] }]
  /transfers/{id}/cancel:
    post:
      tags: [Transfers]
      security: [{ bearerAuth: [] }]
  /payees:
    get:
      tags: [Payees]
      security: [{ bearerAuth: [] }]
    post:
      tags: [Payees]
      security: [{ bearerAuth: [] }]
  /payees/{id}:
    patch:
      tags: [Payees]
      security: [{ bearerAuth: [] }]
    delete:
      tags: [Payees]
      security: [{ bearerAuth: [] }]
...
  /bills/pay:
    post:
      tags: [Bills]
      security: [{ bearerAuth: [] }]
...
  /loans:
    get:
      tags: [Loans]
      security: [{ bearerAuth: [] }]
  /loans/apply:
    post:
      tags: [Loans]
      security: [{ bearerAuth: [] }]
  /loans/{id}/payment:
    post:
      tags: [Loans]
      security: [{ bearerAuth: [] }]
  /loans/{id}/amortization:
    get:
      tags: [Loans]
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT